+++ code/AffableBean/AffableBean/src/java/database/entity/controller/CustomerJpaController.java
@@ -6,8 +6,8 @@
 package database.entity.controller;
 
 import database.entity.Customer;
+import database.entity.controller.exceptions.IllegalOrphanException;
 import database.entity.controller.exceptions.NonexistentEntityException;
-import database.entity.controller.exceptions.PreexistingEntityException;
 import database.entity.controller.exceptions.RollbackFailureException;
 import java.util.List;
 import javax.annotation.Resource;
@@ -16,6 +16,9 @@ import javax.persistence.EntityManagerFactory;
 import javax.persistence.PersistenceUnit;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
+import database.entity.CustomerOrder;
+import java.util.ArrayList;
+import java.util.Collection;
 import javax.transaction.UserTransaction;
 
 /**
@@ -32,21 +35,36 @@ public class CustomerJpaController {
         return emf.createEntityManager();
     }
 
-    public void create(Customer customer) throws PreexistingEntityException, RollbackFailureException, Exception {
+    public void create(Customer customer) throws RollbackFailureException, Exception {
+        if (customer.getCustomerOrderCollection() == null) {
+            customer.setCustomerOrderCollection(new ArrayList<CustomerOrder>());
+        }
         EntityManager em = null;
         try {
             utx.begin();
             em = getEntityManager();
+            Collection<CustomerOrder> attachedCustomerOrderCollection = new ArrayList<CustomerOrder>();
+            for (CustomerOrder customerOrderCollectionCustomerOrderToAttach : customer.getCustomerOrderCollection()) {
+                customerOrderCollectionCustomerOrderToAttach = em.getReference(customerOrderCollectionCustomerOrderToAttach.getClass(), customerOrderCollectionCustomerOrderToAttach.getOrderId());
+                attachedCustomerOrderCollection.add(customerOrderCollectionCustomerOrderToAttach);
+            }
+            customer.setCustomerOrderCollection(attachedCustomerOrderCollection);
             em.persist(customer);
+            for (CustomerOrder customerOrderCollectionCustomerOrder : customer.getCustomerOrderCollection()) {
+                Customer oldCustomerIdOfCustomerOrderCollectionCustomerOrder = customerOrderCollectionCustomerOrder.getCustomerId();
+                customerOrderCollectionCustomerOrder.setCustomerId(customer);
+                customerOrderCollectionCustomerOrder = em.merge(customerOrderCollectionCustomerOrder);
+                if (oldCustomerIdOfCustomerOrderCollectionCustomerOrder != null) {
+                    oldCustomerIdOfCustomerOrderCollectionCustomerOrder.getCustomerOrderCollection().remove(customerOrderCollectionCustomerOrder);
+                    oldCustomerIdOfCustomerOrderCollectionCustomerOrder = em.merge(oldCustomerIdOfCustomerOrderCollectionCustomerOrder);
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
                 utx.rollback();
             } catch (Exception re) {
                 throw new RollbackFailureException("An error occurred attempting to roll back the transaction.", re);
-            }
-            if (findCustomer(customer.getId()) != null) {
-                throw new PreexistingEntityException("Customer " + customer + " already exists.", ex);
             }
             throw ex;
         } finally {
@@ -56,12 +74,45 @@ public class CustomerJpaController {
         }
     }
 
-    public void edit(Customer customer) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void edit(Customer customer) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             utx.begin();
             em = getEntityManager();
+            Customer persistentCustomer = em.find(Customer.class, customer.getCustomerId());
+            Collection<CustomerOrder> customerOrderCollectionOld = persistentCustomer.getCustomerOrderCollection();
+            Collection<CustomerOrder> customerOrderCollectionNew = customer.getCustomerOrderCollection();
+            List<String> illegalOrphanMessages = null;
+            for (CustomerOrder customerOrderCollectionOldCustomerOrder : customerOrderCollectionOld) {
+                if (!customerOrderCollectionNew.contains(customerOrderCollectionOldCustomerOrder)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain CustomerOrder " + customerOrderCollectionOldCustomerOrder + " since its customerId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<CustomerOrder> attachedCustomerOrderCollectionNew = new ArrayList<CustomerOrder>();
+            for (CustomerOrder customerOrderCollectionNewCustomerOrderToAttach : customerOrderCollectionNew) {
+                customerOrderCollectionNewCustomerOrderToAttach = em.getReference(customerOrderCollectionNewCustomerOrderToAttach.getClass(), customerOrderCollectionNewCustomerOrderToAttach.getOrderId());
+                attachedCustomerOrderCollectionNew.add(customerOrderCollectionNewCustomerOrderToAttach);
+            }
+            customerOrderCollectionNew = attachedCustomerOrderCollectionNew;
+            customer.setCustomerOrderCollection(customerOrderCollectionNew);
             customer = em.merge(customer);
+            for (CustomerOrder customerOrderCollectionNewCustomerOrder : customerOrderCollectionNew) {
+                if (!customerOrderCollectionOld.contains(customerOrderCollectionNewCustomerOrder)) {
+                    Customer oldCustomerIdOfCustomerOrderCollectionNewCustomerOrder = customerOrderCollectionNewCustomerOrder.getCustomerId();
+                    customerOrderCollectionNewCustomerOrder.setCustomerId(customer);
+                    customerOrderCollectionNewCustomerOrder = em.merge(customerOrderCollectionNewCustomerOrder);
+                    if (oldCustomerIdOfCustomerOrderCollectionNewCustomerOrder != null && !oldCustomerIdOfCustomerOrderCollectionNewCustomerOrder.equals(customer)) {
+                        oldCustomerIdOfCustomerOrderCollectionNewCustomerOrder.getCustomerOrderCollection().remove(customerOrderCollectionNewCustomerOrder);
+                        oldCustomerIdOfCustomerOrderCollectionNewCustomerOrder = em.merge(oldCustomerIdOfCustomerOrderCollectionNewCustomerOrder);
+                    }
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -71,7 +122,7 @@ public class CustomerJpaController {
             }
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                String id = customer.getId();
+                Short id = customer.getCustomerId();
                 if (findCustomer(id) == null) {
                     throw new NonexistentEntityException("The customer with id " + id + " no longer exists.");
                 }
@@ -84,7 +135,7 @@ public class CustomerJpaController {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void destroy(Short id) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             utx.begin();
@@ -92,9 +143,20 @@ public class CustomerJpaController {
             Customer customer;
             try {
                 customer = em.getReference(Customer.class, id);
-                customer.getId();
+                customer.getCustomerId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The customer with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<CustomerOrder> customerOrderCollectionOrphanCheck = customer.getCustomerOrderCollection();
+            for (CustomerOrder customerOrderCollectionOrphanCheckCustomerOrder : customerOrderCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Customer (" + customer + ") cannot be destroyed since the CustomerOrder " + customerOrderCollectionOrphanCheckCustomerOrder + " in its customerOrderCollection field has a non-nullable customerId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(customer);
             utx.commit();
@@ -134,7 +196,7 @@ public class CustomerJpaController {
         }
     }
 
-    public Customer findCustomer(String id) {
+    public Customer findCustomer(Short id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Customer.class, id);
